pyproject.toml
[project]
authors = [
    {name = "pabl-o-ce", email = "cye@poscye.com"},
    {name = "Maximilian-Winter", email = "maximilian.winter.91@gmail.com"}
]
name = "dojo-bot"
version = "0.1.0"
description = "Dojo bot: tst tools"
license = {file = "LICENSE.txt"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "discord-py-interactions ==5.14.0",
    "python-dotenv ==1.0.1",
    "redis ==5.0.4",
    "discord-py ==2.3.2",
]

[project.scripts]
start-bot = "main:main"

[project.optional-dependencies]
dev = [
    "pylint ==3.2.3"
]

[build-system]
requires = [
    "packaging",
    "setuptools",
    "wheel",
]
build-backend = "setuptools.build_meta"
